Got a rudimentary version of cloud storage working.
Error handling is largely missing.  Also, image loading is happening directly via the cloud storage URL.  Once data is encrypted we'll need to fetch, then decrypt, and then use a local URL for the image data.  Downloading works using the current method, but it will not work in the future.