Refactored store image logic to better use network resources.
Previously we'd not make use of the initial file upload URL but instead wait for the DB to store the data, fetch its data, and serve it on a new local URL.  This led to very slow uploads.

The new approach is a bit sensitive in that the UI will for some period of time show an incorrect tag list while the file is being uploaded in the background.