Experimental approach with batching DB reads when loading an image-gallery with many image-cards.
This helps, but does not alleviate issues with "All" where the initial request to the DB blocks for a long time.  There is a strange 3-4 second cadence on loading images, and the number of images that are loaded simultaneously don't really matter from 5-25 there is this "pulse" of resolving image data and showing thumbnails.

I also updated the logic of subscribing to image data to lessen the load on the DB -- once data is fetched, there are no more updates to have, so the image-card now unsubscribes immediately after receiving the first ImageData.